+++ somni-expr/src/visitor.rs
@@ -339,7 +339,6 @@ where
                 }
             } else {
                 // Reset result if we have statements after implicit returns.
-                // FIXME: This is a hack to handle return values from blocks that shouldn't be returned.
                 body_result = StatementResult::EndOfBody;
             }
         }